server/requirements.txt
Flask
flask-cors
google-cloud-bigquery